pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whoop-mcp"
version = "0.1.0"
description = "MCP server for the WHOOP API — access recovery, sleep, workout, and cycle data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/anaselmhamdi/whoop-mcp"
Repository = "https://github.com/anaselmhamdi/whoop-mcp"

[project.scripts]
whoop-mcp = "whoop_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
